+++ packages/datasource-toolkit/CHANGELOG.md
@@ -1,3 +1,10 @@
+# @forestadmin/datasource-toolkit [1.5.0](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-toolkit@1.4.3...@forestadmin/datasource-toolkit@1.5.0) (2023-04-26)
+
+
+### Features
+
+* **datasource-customizer:** add support for binary fields ([#673](https://github.com/ForestAdmin/agent-nodejs/issues/673)) ([419727c](https://github.com/ForestAdmin/agent-nodejs/commit/419727cce812af10fa5917f1b94a2064d06883f7))
+
 ## @forestadmin/datasource-toolkit [1.4.3](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-toolkit@1.4.2...@forestadmin/datasource-toolkit@1.4.3) (2023-04-19)