+++ packages/forestadmin-client/CHANGELOG.md
@@ -1,3 +1,13 @@
+## @forestadmin/forestadmin-client [1.1.13](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/forestadmin-client@1.1.12...@forestadmin/forestadmin-client@1.1.13) (2023-04-26)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-toolkit:** upgraded to 1.5.0
+
 # @forestadmin/forestadmin-client [1.2.0-alpha-webhook-custom-actions.2](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/forestadmin-client@1.2.0-alpha-webhook-custom-actions.1...@forestadmin/forestadmin-client@1.2.0-alpha-webhook-custom-actions.2) (2023-04-24)
 
 # @forestadmin/forestadmin-client [1.2.0-alpha-webhook-custom-actions.1](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/forestadmin-client@1.1.12...@forestadmin/forestadmin-client@1.2.0-alpha-webhook-custom-actions.1) (2023-04-20)