+++ packages/plugin-flattener/CHANGELOG.md
@@ -1,3 +1,24 @@
+## @forestadmin/plugin-flattener [1.0.19](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-flattener@1.0.18...@forestadmin/plugin-flattener@1.0.19) (2023-04-27)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.7.1
+
+## @forestadmin/plugin-flattener [1.0.18](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-flattener@1.0.17...@forestadmin/plugin-flattener@1.0.18) (2023-04-26)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-toolkit:** upgraded to 1.5.0
+* **@forestadmin/datasource-customizer:** upgraded to 1.7.0
+
 ## @forestadmin/plugin-flattener [1.0.18-alpha-webhook-custom-actions.1](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-flattener@1.0.17...@forestadmin/plugin-flattener@1.0.18-alpha-webhook-custom-actions.1) (2023-04-20)